Split def for working with files in another pyFile
